% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macca_plot.R
\name{macca_inflation_plot}
\alias{macca_inflation_plot}
\title{Generate an interactive plot of inflation rates of all countries from 2011-2020}
\usage{
macca_inflation_plot(inflachoice)
}
\arguments{
\item{inflachoice}{A character string indicating the type of inflation to plot. It can be "usd",
"us dollar", "us", or "local price". If it matches any of these options (case-insensitive),
the plot will show the inflation rates based on the Dollar price, otherwise it will show
the inflation rates based on the Local price.}
}
\value{
An interactive plotly object displaying the inflation rates over time.
}
\description{
This function generates an interactive plot showing the inflation rates over time for different countries.
}
\examples{
macca_inflation_plot("usd")
macca_inflation_plot("Local Price")

}
