% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maccaindex-package.R
\docType{package}
\name{maccaindex-package}
\alias{maccaindex}
\alias{maccaindex-package}
\title{maccaindex: Visualization of Big Index Data And Miscellaneous Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains the 1) Raw data of Big Mac Index from 2011 to 2020; 2) A Shiny App which visualizes this dataset to ignite visual learning, 3) Customizable plotting function which takes in various inputs and display accordingly; and 4) A function can take in either one or two arguments and return a dataframe dependent on the types of the inputs (and how many imputs were called).
}
\author{
\strong{Maintainer}: Minh Khanh Pham \email{mpham0054@student.monash.edu} (33602875)

}
\keyword{internal}
