% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_maccapp.R
\name{maccapp}
\alias{maccapp}
\title{Run the Big Mac Shiny App}
\usage{
maccapp()
}
\value{
A Shiny App Interface
}
\description{
This function launches the Big Mac Shiny app.
}
\details{
maccapp
}
\examples{
\dontrun{
maccapp()
}

}
