% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_average.R
\name{magicwand_df}
\alias{magicwand_df}
\title{Magic Wand DataFrame}
\usage{
magicwand_df(vectorx, vectory = NULL)
}
\arguments{
\item{vectorx}{A vector of values (either a numerical, character, factor).}

\item{vectory}{A vector of values (optional).}
}
\value{
A customized data frame based on the input vectors.
\itemize{
\item If \code{vectorx} is a numerical vector of at least 5 values, returns a data frame with two columns:
\itemize{
\item Names: the names of the values (Minimum, Lower Hinge, Q1, Median, Q3, Upper Hinge, Maximum)
\item Values: the corresponding summary values
}
\item If \code{vectorx} is a character or factor vector, returns a data frame with two columns:
\itemize{
\item Values: the unique values in \code{vectorx}
\item Frequencies: the frequencies of each value
}
\item If both \code{vectorx} and \code{vectory} are numerical vectors of the same size, returns a data frame with one column:
\itemize{
\item Coefficient: the correlation coefficient between \code{vectorx} and \code{vectory}
}
\item If one of \code{vectorx} or \code{vectory} is numerical and the other is a character or factor vector, returns a data frame with two columns:
\itemize{
\item Levels: the unique levels in the character/factor vector
\item Mean: the mean value of the numerical vector for each level
}
\item If both \code{vectorx} and \code{vectory} are character/factor vectors, returns a data frame representing the cross-tabulation of \code{vectorx} and \code{vectory}
}
}
\description{
A function that generates a customized data frame based on the input vectors.
}
\examples{

# Single argument functionality for numerical vector
vectorx <- c(1, 2, 3, 4, 5)
result <- magicwand_df(vectorx)
print(result)

# Single argument functionality for character/factor vector
vectorx <- c("apple", "banana", "apple", "orange", "banana", "orange")
result <- magicwand_df(vectorx)
print(result)

# Two-argument functionality for numerical vectors
vectorx <- c(1, 2, 3)
vectory <- c(4, 5, 6)
result <- magicwand_df(vectorx, vectory)
print(result)

# Two-argument functionality for numerical and character/factor vectors
vectorx <- c(1, 2, 1, 2, 1)
vectory <- c("apple", "banana", "apple", "banana", "banana")
result <- magicwand_df(vectorx, vectory)
print(result)

# Two-argument functionality for character/factor vectors
vectorx <- c("apple", "banana", "apple", "orange", "banana")
vectory <- c("red", "yellow", "red", "yellow", "red")
result <- magicwand_df(vectorx, vectory)
print(result)

}
